refactor: switch to resource endpoint
