+++ common/src/models/mod.rs
@@ -1,6 +1,7 @@
 pub mod deployment;
 pub mod error;
 pub mod project;
+pub mod resource;
 pub mod secret;
 pub mod service;
 pub mod stats;
