+++ resources/shared-db/src/postgres.rs
@@ -1,8 +1,8 @@
 use async_trait::async_trait;
 use serde::Serialize;
-use shuttle_service::{database, error::CustomError, Error, Factory, ResourceBuilder, Type};
-
-use crate::SharedDbOutput;
+use shuttle_service::{
+    database, error::CustomError, DbOutput, Error, Factory, ResourceBuilder, Type,
+};
 
 #[derive(Serialize)]
 pub struct Postgres {
@@ -14,7 +14,7 @@ pub struct Postgres {
 impl ResourceBuilder<sqlx::PgPool> for Postgres {
     const TYPE: Type = Type::Database(database::Type::Shared(database::SharedEngine::Postgres));
 
-    type Output = SharedDbOutput;
+    type Output = DbOutput;
 
     fn new() -> Self {
         Self { local_uri: None }
@@ -22,16 +22,16 @@ impl ResourceBuilder<sqlx::PgPool> for Postgres {
 
     async fn output(self, factory: &mut dyn Factory) -> Result<Self::Output, Error> {
         let info = match factory.get_environment() {
-            shuttle_service::Environment::Production => SharedDbOutput::Shared(
+            shuttle_service::Environment::Production => DbOutput::Info(
                 factory
                     .get_db_connection(database::Type::Shared(database::SharedEngine::Postgres))
                     .await?,
             ),
             shuttle_service::Environment::Local => {
                 if let Some(local_uri) = self.local_uri {
-                    SharedDbOutput::Local(local_uri)
+                    DbOutput::Local(local_uri)
                 } else {
-                    SharedDbOutput::Shared(
+                    DbOutput::Info(
                         factory
                             .get_db_connection(database::Type::Shared(
                                 database::SharedEngine::Postgres,
@@ -47,8 +47,8 @@ impl ResourceBuilder<sqlx::PgPool> for Postgres {
 
     async fn build(build_data: &Self::Output) -> Result<sqlx::PgPool, Error> {
         let connection_string = match build_data {
-            SharedDbOutput::Local(local_uri) => local_uri.clone(),
-            SharedDbOutput::Shared(info) => info.connection_string_private(),
+            DbOutput::Local(local_uri) => local_uri.clone(),
+            DbOutput::Info(info) => info.connection_string_private(),
         };
 
         let pool = sqlx::postgres::PgPoolOptions::new()
